24/2/8 ia-update code restructure
tODO: contract_list:
age - xx.0
