+++ cl_selenium/cl_investment.py
@@ -2,14 +2,13 @@ import sys
 import os
 from datetime import datetime
 
-from cl_scrap import wd
 from selenium import webdriver
 from selenium.webdriver.chrome import webdriver
 from selenium.webdriver.common.by import By
 
 from cl_selenium import cl_selectors
 
-from cl_scrap import cl_account, driver_setup, login
+from cl_scrap import cl_account, login
 paths = cl_selectors.fund_paths()
 statement_date = wd.find_element(By.XPATH, paths['statement_date']).text
 formatted_date = datetime.strptime(statement_date, '%b. %d, %Y').strftime('%Y-%m-%d')
