+++ ia_app.py
@@ -1,26 +1,14 @@
 import os
-import sys
 import threading
-from datetime import datetime
 
-import pandas as pd
-
-from ia_selenium import ia_scrap, keys
-from ia_selenium.ia_scrap import get_control, driver_setup, ia_app, create_table, ia_threading, merge_tables
+from ia_selenium import ia_scrap
+from ia_selenium.ia_contract_list import click_contract_list
+from ia_selenium.ia_scrap import ia_threading
 from ia_selenium.split_excel import split_excel
+from utilities.get_confs import get_confs
+from utilities.save_csv import get_csv_file_names, save_table_into_csv
+from utilities.tables_utilities import merge_tables
 
-# ia_wd : chrome driver
-# maximum_iteration: max iteration of scrap loop
-# control_unit: get control unit for scrap
-# ia_parameters: get parameters setting
-# tables: information containers
-# csvs: the file location
-
-
-# csvs
-# tables
-# control_unit
-# set up the control unit & recovery_file unit
 # {
 #         'csvs': csvs,
 #         'parameters': ia_parameters,
@@ -31,10 +19,10 @@ from ia_selenium.split_excel import split_excel
 #         'threading_tables': threading_tables,
 #         'thread_number': thread_number
 #  }
-confs = ia_scrap.ia_get_confs()
+confs = get_confs('ia')
 
-contract_files = split_excel(os.path.join(confs['parameters']['csv_path'], confs['parameters']['contracts'], confs['contract_file']),
-                             os.path.join(confs['parameters']['csv_path'], confs['date_today']),
+contract_files = split_excel(confs['contract_path'],
+                             confs['csvs'],
                              confs['thread_number'])
 
 threads_list = []
@@ -53,13 +41,13 @@ for t in threads_list:
 tables = merge_tables(confs)
 
 # record file names
-files = ia_scrap.get_csv_file_names(confs['csvs'])
+files = get_csv_file_names(confs['csvs'])
 
 # save tables into csv files
-ia_scrap.save_table_into_csv(confs['control_unit'], tables, files)
+save_table_into_csv(confs['control_unit'], tables, files)
 
 # save csv files into db
 ia_scrap.save_csv_to_db(confs['control_unit'], files, tables)
 
 # request contract numbers
-ia_scrap.click_contract_list(confs)
+click_contract_list(confs)
